👷 Refactor GitHub Action to comment docs deployment URLs and update token, preparing for GitHub org
