+++ .github/workflows/deploy-docs.yml
@@ -6,6 +6,11 @@ on:
     types:
       - completed
 
+permissions:
+  deployments: write
+  issues: write
+  pull-requests: write
+
 jobs:
   deploy-docs:
     runs-on: ubuntu-latest
@@ -22,7 +27,7 @@ jobs:
       - uses: actions/download-artifact@v4
         with:
           path: ./site/
-          pattern: docs-site
+          pattern: docs-site-*
           merge-multiple: true
           github-token: ${{ secrets.GITHUB_TOKEN }}
           run-id: ${{ github.event.workflow_run.id }}
@@ -34,13 +39,26 @@ jobs:
         with:
           apiToken: ${{ secrets.CLOUDFLARE_API_TOKEN }}
           accountId: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
-          projectName: typertiangolo
+          projectName: fastapitiangolo
           directory: './site'
           gitHubToken: ${{ secrets.GITHUB_TOKEN }}
           branch: ${{ ( github.event.workflow_run.head_repository.full_name == github.repository && github.event.workflow_run.head_branch == 'master' && 'main' ) || ( github.event.workflow_run.head_sha ) }}
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+      - uses: actions/cache@v4
+        id: cache
+        with:
+          path: ${{ env.pythonLocation }}
+          key: ${{ runner.os }}-python-github-actions-${{ env.pythonLocation }}-${{ hashFiles('requirements-github-actions.txt') }}-v01
+      - name: Install GitHub Actions dependencies
+        if: steps.cache.outputs.cache-hit != 'true'
+        run: pip install -r requirements-github-actions.txt
       - name: Comment Deploy
         if: steps.deploy.outputs.url != ''
-        uses: ./.github/actions/comment-docs-preview-in-pr
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-          deploy_url: "${{ steps.deploy.outputs.url }}"
+        run: python ./scripts/comment_docs_deploy_url_in_pr.py
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          DEPLOY_URL: ${{ steps.deploy.outputs.url }}
+          COMMIT_SHA: ${{ github.event.workflow_run.head_sha }}
