+++ requirements-github-actions.txt
@@ -1,0 +1,4 @@
+PyGithub>=2.3.0,<3.0.0
+pydantic>=2.5.3,<3.0.0
+pydantic-settings>=2.1.0,<3.0.0
+httpx>=0.27.0,<0.28.0
